refactor: 예약 관련 검증 로직 Service로 이동

+++ src/main/java/roomescape/service/ReservationService.java
@@ -18,6 +18,8 @@ public class ReservationService {
 
     @Transactional
     public Reservation addReservation(ReservationRequestDto request) {
+        validate(request);
+
         Time time = timeService.findById(request.time());
         Reservation reservation = Reservation.builder()
                 .name(request.name())
@@ -31,11 +33,39 @@ public class ReservationService {
 
     @Transactional
     public void deleteReservation(Long reservationId) {
+        if (reservationDao.findById(reservationId) == null) {
+            throw new IllegalArgumentException("해당 예약이 존재하지 않습니다");
+        }
+
         reservationDao.delete(reservationId);
     }
 
     @Transactional
     public List<Reservation> findAll() {
         return reservationDao.findAll();
+    }
+
+    private void validate(ReservationRequestDto request) {
+        validateName(request.name());
+        validateDate(request.date());
+        validateTime(request.time());
+    }
+
+    private void validateName(String name) {
+        if (name == null || name.isEmpty()) {
+            throw new IllegalArgumentException("이름을 입력해주세요");
+        }
+    }
+
+    private void validateDate(String date) {
+        if (date == null || date.isEmpty()) {
+            throw new IllegalArgumentException("날짜를 입력해주세요");
+        }
+    }
+
+    private void validateTime(Long time) {
+        if (time == null) {
+            throw new IllegalArgumentException("시간을 입력해주세요");
+        }
     }
 }
